refactor(dsp): 重构频谱与滤波模块，新增配置与频谱输出
1. 重构采样率类型为float统一处理滤波逻辑
2. 新增WindowType枚举统一窗函数管理
3. 添加频谱分析配置项与信号输出
4. 优化FFT处理逻辑与代码结构

+++ realtime_app/dsp/filter_strategy.py
@@ -49,7 +49,7 @@ def get_strategy() -> FilterStrategy:
 
 def apply_filters(
     data: np.ndarray,
-    sampling_rate: int = 250,
+    sampling_rate: float = 250.0,
     highpass: float = 0.5,
     lowpass: float = 45.0,
     order: int = 4,
